crypto_analyzer/requirements.txt
# Crypto Analyzer Dependencies
aiohttp>=3.8.0
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
